api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==1.3.5
numpy==1.24.3
FlagEmbedding==1.1.7
tqdm==4.66.1
pydantic==2.4.2
starlette==0.27.0 